requirements.txt
fasttext==0.9.2
nltk==3.8.1
numpy==1.24.3
pandas==1.5.3
pyTelegramBotAPI==4.11.0
python_docx==0.8.11
streamlit==1.21.0
telebot~=0.0.5